small performance optimizations
